% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adea_setup.R
\name{adea_setup}
\alias{adea_setup}
\title{Standarize input and output}
\usage{
adea_setup(input, output, ux = NULL, vy = NULL)
}
\arguments{
\item{input}{A matrix or a data frame containing the inputs of the units to be evaluated, with one row for each DMU and one column for each input.}

\item{output}{A matrix or a data frame containing the outputs of the units to be evaluated, with one row for each DMU and one column for each output.}
}
\value{
a list with input and output converted
}
\description{
Input and outputs are converted, if require, into a matrix.
The names of the columns and rows are also set in both.
When input and output arguments do not provide names, standard ones are given.
This functions does not check the arguments, so call the adea_check function first if necessary.
}
\keyword{internal}
