% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardealers1.R
\docType{data}
\name{cardealers1}
\alias{cardealers1}
\title{A data set about car dealers, 1 of 4, to be used in DEA}
\format{
\code{Employees}: The number of employees in the car dealer to be considered as input variable in DEA.

\code{Sold}: The sold amount by the car dealer to be considered as output variable in DEA.
}
\source{
\emph{Vicente Coll Serrano y Olga M. Blasco Blasco} (2006) 
Evaluacion de la eficiencia mediante el analisis envolvente de datos. 
Electronic edition.
\url{https://www.eumed.net/libros-gratis/2006c/197/}
}
\description{
An example dataset with six car dealers data to be use in DEA (Data Envelopment Analysis).
}
\details{
The purpose of this dataframe is to be used in DEA (Data Envelopment Analysis).
DEA rate each unit, in this case car dealers, in terms of efficiency.
Less input and more output lead to higher efficiency scores.
}
\section{Usage}{

data('cardealers1')
}

\seealso{
See also \code{\link{adea-package}}.
}
\keyword{data}
