% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardealers4.R
\docType{data}
\name{cardealers4}
\alias{cardealers4}
\title{A data set about car dealers, 4 of 4, to be used in DEA}
\format{
\code{Employees}: Number of employees to be considered as input variable in DEA.

\code{Depreciation}: Depreciation of stored to be considered as input variable in DEA.

\code{Sold}: Number of cars sold as output variable in DEA.

\code{Orders}: Number of work orders in garage as output variable in DEA.
}
\source{
\emph{Vicente Coll Serrano y Olga M. Blasco Blasco} (2006) 
Evaluacion de la eficiencia mediante el analisis envolvente de datos. 
Electronic edition.
\url{https://www.eumed.net/libros-gratis/2006c/197/}
}
\description{
An example dataset with six car dealers to be use in DEA (Data envelopment analysis).
}
\details{
The purpose of this dataframe is to be used in DEA (Data Envelopment Analysis).
DEA rate each unit, in this case car dealers, in terms of efficiency.
Less input and more output lead to higher efficiency scores.
}
\section{Usage}{

data('cardealers4')
}

\seealso{
See also \code{\link{adea-package}}.
}
\keyword{data}
