% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adea.R
\name{print.adea}
\alias{print.adea}
\title{Print method for adea class}
\usage{
\method{print}{adea}(x, ...)
}
\arguments{
\item{x}{Object of class adea to print.}

\item{...}{Optional arguments to "print".}
}
\description{
Print method for adea class
}
