% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adeahierarchical.R
\name{print.adeahierarchical}
\alias{print.adeahierarchical}
\title{Print method for adeahierarchical class}
\usage{
\method{print}{adeahierarchical}(x, ...)
}
\arguments{
\item{x}{Object of class adeahierarchical to print.}

\item{...}{Optional arguments to "print".}
}
\description{
For the final model of adea_hierarchical function prints a summary, the input variables, and outputs variables.
}
