% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adealoadleverage.R
\name{print.adealoadleverage}
\alias{print.adealoadleverage}
\title{Print method for adealoadleverage class}
\usage{
\method{print}{adealoadleverage}(x, ...)
}
\arguments{
\item{x}{Object of class adealoadleverage to print.}

\item{...}{Optional arguments to "print".}
}
\description{
Print adealoadleverage results in human readable way
}
