% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adeaparametric.R
\name{print.adeaparametric}
\alias{print.adeaparametric}
\title{Print method for adeaparametric class}
\usage{
\method{print}{adeaparametric}(x, ...)
}
\arguments{
\item{x}{Object of class adeaparametric to print.}

\item{...}{Optional arguments to "print".}
}
\description{
For the final model of adea_parametric function prints a summary, the input variables, and outputs variables.
}
