% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adeastepwise.R
\name{print.adeastepwise}
\alias{print.adeastepwise}
\title{Print method for adeastepwise class}
\usage{
\method{print}{adeastepwise}(x, ...)
}
\arguments{
\item{x}{Object of class stepwise to print.}

\item{...}{Optional arguments to "print".}
}
\description{
For the final model of adeastepwise function prints a summary, the input variables, and outputs variables.
}
