% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cadea.R
\name{print.cadea}
\alias{print.cadea}
\title{Print method for cadea class}
\usage{
\method{print}{cadea}(x, ...)
}
\arguments{
\item{x}{Object of class cadea to print.}

\item{...}{Optional arguments to "print".}
}
\description{
Print method for cadea class
}
