% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dea.R
\name{print.dea}
\alias{print.dea}
\title{Print method for dea class}
\usage{
\method{print}{dea}(x, ...)
}
\arguments{
\item{x}{Object of class dea to print.}

\item{...}{Optional arguments to "print".}
}
\description{
Print method for dea class
}
