% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fsdea.R
\name{print.fsdea}
\alias{print.fsdea}
\title{Print method for fsdea class}
\usage{
\method{print}{fsdea}(x, ...)
}
\arguments{
\item{x}{Object of class fsdea to print.}

\item{...}{Optional arguments to "print".}
}
\description{
Print method for fsdea class
}
