% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adeaparametric.R
\name{summary.adeaparametric}
\alias{summary.adeaparametric}
\title{Summary method for adeaparametric class}
\usage{
\method{summary}{adeaparametric}(object, ...)
}
\arguments{
\item{object}{is the object of class adeaparametric to summarise}

\item{...}{For compatibility reason, see note about `eff.tolerance` parameter.}
}
\description{
For the final model of adea_parametric function print the model name, orientation, load orientation, a summary, the input variables, and outputs variables.
}
\details{
The default tolerance to consider a DMU as efficient one is .001 in reports.
Use `eff.tolerance` parameter to consider another tolerance between 0 and 1.
}
