% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cadea.R
\name{summary.cadea}
\alias{summary.cadea}
\title{summary method for cadea class}
\usage{
\method{summary}{cadea}(object, ...)
}
\arguments{
\item{object}{is the object of class cadea to summarise}

\item{...}{optional arguments to 'print'}
}
\description{
summary method for cadea class
}
