% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dea.R
\name{summary.dea}
\alias{summary.dea}
\title{summary method for dea class}
\usage{
\method{summary}{dea}(object, ...)
}
\arguments{
\item{object}{is the object of class dea to summarise}

\item{...}{For compatibility reason, see note about `eff.tolerance` parameter.}
}
\value{
This function returns the following information, as a named list:
\itemize{
  \item Model name: if it is previously set
  \item Orientation: Model orientation
  \item Inputs: Names of the input variables in the model
  \item Outputs: Names of the output variables in the model
  \item nInputs: Number of the input variables in the model
  \item nOutputs: Number of the output variables in the model
  \item nVariables: Total number of input and output variables in the model
  \item n: Number of DMUs
  \item nEfficients: Number of efficient DMUs taking into account whithin eff.tolerance tolerance.
  \item Mean: Mean of efficiency scores.
  \item sd: Standard deviation of efficiency scores.
  \item Min.: Minimun value of efficiency scores.
  \item 1st Qu.: First quater of efficiency scores.
  \item Median: Median of efficiency scores.
  \item 3rd Qu.: Third quater of efficiency scores.
  \item Max.: Maximum value of efficiency scores.
}
}
\description{
Extract information from a dea class object and compute some more.
}
\details{
The default tolerance to consider a DMU as efficient one is .001 in reports.
Use `eff.tolerance` parameter to consider another tolerance between 0 and 1.
}
