% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fsdea.R
\name{summary.fsdea}
\alias{summary.fsdea}
\title{summary method for fsdea class}
\usage{
\method{summary}{fsdea}(object, ...)
}
\arguments{
\item{object}{is the object of class dea to summarise}

\item{...}{For compatibility reason, see note about `eff.tolerance` parameter.}
}
\value{
This function returns the following information, as a named vector:
\itemize{
  \item Model name: if it is previously set
  \item Orientation: Model orientation
  \item Inputs: Names of the input variables in the model
  \item Outputs: Names of the output variables in the model
  \item n: Number of DMUs
  \item nEfficients: Number of efficient DMUs taking into account whithin eff.tolerance tolerance.
  \item Mean: This and following statistical measure of the efficiency scores. Its mean.
  \item sd: Its standard deviation.
  \item Min.: Its minimun value.
  \item 1st Qu.: Its first quater.
  \item Median: Its median.
  \item 3rd Qu.: Its third quater.
  \item Max.: Its maximum value.
  \item nInputs: Number of input variables to select
  \item nOutputs: Number of output variables to select
  \item nTotal: Total number of variables to select
  \item iSelected: Names of the selected input variables
  \item oSelected: Names of the selected output variables
}
}
\description{
Extract information from a fsdea class object and compute some more.
The default tolerance to consider a DMU as efficient one is .001 in reports.
Use `eff.tolerance` parameter to consider another tolerance between 0 and 1.
}
