% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokyo_libraries.R
\docType{data}
\name{tokyo_libraries}
\alias{tokyo_libraries}
\title{A data set of Tokyo libraries}
\format{
\code{Area.I1}: Area of library.

\code{Books.I2}: Number of books.

\code{Staff.I3}: Staff members.

\code{Populations.I4}: Population of influence area.

\code{Regist.O1}: Registered people.

\code{Borrow.O2}: Borrowed books.
}
\description{
\code{tokyo_libraries} involves a set of 23 libraries in Tokyo and it has been used frequently in DEA literature.
}
\details{
The purpose of this dataframe is to be used in DEA (Data Envelopment Analysis).

The input variables are \code{Area.I1}, \code{Books.I2}, \code{Staff.I3} and \code{Populations.I4}.

The output variables are \code{Regist.O1} and \code{Borrow.O2}.
}
\section{Usage}{

data('tokyo_libraries')
}

\seealso{
See also \code{\link{adea-package}}.
}
\keyword{data}
