% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_dea.R
\name{virtual_dea}
\alias{virtual_dea}
\title{Compute virtual input and output for (standard) DEA model}
\usage{
virtual_dea(input, output, orientation, ux, vy)
}
\arguments{
\item{input}{A matrix or a data frame containing the inputs of the units to be evaluated, with one row for each DMU and one column for each input.}

\item{output}{A matrix or a data frame containing the outputs of the units to be evaluated, with one row for each DMU and one column for each output.}

\item{orientation}{Use "input" for input orientation or "output" for output orientation in DEA model.
The default is "input".}
}
\value{
{vinput, voutput}
}
\description{
For the given orientation, input, output and weights compute virtual input and output as described in @references A new approach to the bi-dimensional representation of the DEA efficient frontier with multiple inputs and outputscompute DMU's efficiencies
}
\keyword{internal}
