% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adea_setup.R
\name{adea_setup}
\alias{adea_setup}
\title{Convert input and output in 1 column matrix}
\usage{
adea_setup(input, output)
}
\arguments{
\item{input}{A matrix or a data frame with the inputs of units to be evaluated, one row for each DMU and one column for each input.}

\item{output}{A matrix or a data frame with the outputs of units to be evaluated, one row for each DMU and one column for each output.}
}
\value{
a list with input and output converted
}
\description{
Setup names for DMUs, input and outputs
}
\keyword{internal}
