% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_solve_dea.R
\name{lp_solve_dea}
\alias{lp_solve_dea}
\title{Build dea problem}
\usage{
lp_solve_dea(input, output, orientation = c("input", "output"), solve = FALSE)
}
\arguments{
\item{input}{A matrix or a data frame containing the inputs of the units to be evaluated, with one row for each DMU and one column for each input.}

\item{output}{A matrix or a data frame containing the outputs of the units to be evaluated, with one row for each DMU and one column for each output.}

\item{orientation}{Use "input" for input orientation or "output" for output orientation in DEA model.
The default is "input".}

\item{solve}{If TRUE then solve dea model}
}
\value{
list(lp, status, ux, vy, eff)
}
\description{
For the given input and output build the lp_solve problem and return it.
}
\keyword{internal}
