% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adea.R
\name{summary.adea}
\alias{summary.adea}
\title{summary method for adea class}
\usage{
\method{summary}{adea}(object, ...)
}
\arguments{
\item{object}{is the object of class adea to summarise}

\item{...}{optional arguments to 'print'}
}
\description{
summary method for adea class
}
