% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adeahierarchical.R
\name{summary.adeahierarchical}
\alias{summary.adeahierarchical}
\title{Summary method for adeahierarchical class}
\usage{
\method{summary}{adeahierarchical}(object, ...)
}
\arguments{
\item{object}{Is the object of class adeahierarchical to summarise.}

\item{...}{Optional arguments to "print".}
}
\description{
For the final model of adea_hierarchical function print the model name, orientation, load orientation, a summary, the input variables, and outputs variables.
}
