% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adeaparametric.R
\name{summary.adeaparametric}
\alias{summary.adeaparametric}
\title{summary method for adeaparametric class}
\usage{
\method{summary}{adeaparametric}(object, ...)
}
\arguments{
\item{object}{is the object of class adeaparametric to summarise}

\item{...}{optional arguments to 'print'}
}
\description{
For the final model of adea_parametric function print the model name, orientation, load orientation, a summary, the input variables, and outputs variables.
}
