% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.adeastepwise.R
\name{summary.adeastepwise}
\alias{summary.adeastepwise}
\title{Summary method for adeastepwise class}
\usage{
\method{summary}{adeastepwise}(object, ...)
}
\arguments{
\item{object}{is the object of class adeastepwise to summarise.}

\item{...}{Optional arguments to "print".}
}
\description{
For the final model of adea_stepwise function print the model name, orientation, load orientation, a summary, the input variables, and outputs variables.
}
