% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cadea.R
\name{summary.cadea}
\alias{summary.cadea}
\title{summary method for cadea class}
\usage{
\method{summary}{cadea}(object, ...)
}
\arguments{
\item{object}{is the object of class cadea to summarise}

\item{...}{For compatibility reason, see note about `eff.tolerance` parameter.}
}
\description{
Extract information from a fsdea class object and compute some more.
The default tolerance to consider a DMU as efficient one is .001 in reports.
Use `eff.tolerance` parameter to consider another tolerance between 0 and 1.
}
